azure-storage-connection-string/requirements.txt
azure-storage-blob==12.18.3